src/requirements.txt
boto3==1.26.0
line-bot-sdk==2.4.2
requests==2.28.1